use pip for base-notebook image
this is an initial attempt to get base-notebook image up and running
using pip.

for now I'm trying to get it up using minimal components, some of the
components are missing from the image as of now like pandoc.

I don't intend to include that because I don't use pdf functionality.

Signed-off-by: kranurag7 <81210977+kranurag7@users.noreply.github.com>